Current versions of files for flashing
clear_flash.hex is a firmware and keymap from EasyAVR which fixed an issue where I made a board unusuable with a development branch of qmk-firmware. It somehow damaged the keymap in a way I couldn't easily fix.

satan.hex is what I am currently flashing to RGB60 boards and it works perfectly.